requirements.txt
boto==2.46.1
dj-database-url==0.4.2
Django==1.11
django-boto==0.3.11
djangorestframework==3.6.2
gunicorn==19.7.1
ipdb==0.10.2
ipython==6.0.0
isort==4.2.5
psycopg2==2.7.1
pylint==1.7.1
whitenoise==3.3.0
